moved login to components and added to README
